Split client command api into its own module
