+++ library/command/client_command/src/main/java/org/quiltmc/qsl/command/api/client/ClientCommandRegistrationCallback.java
