+++ library/command/command/src/testmod/resources/fabric.mod.json
@@ -20,9 +20,6 @@
   "entrypoints": {
     "main": [
       "org.quiltmc.qsl.command.test.CommandApiTest"
-    ],
-    "client": [
-      "org.quiltmc.qsl.command.test.client.CommandApiTestClient"
     ]
   },
   "description": "Testmod for Quilt APIs relating to commands.",
